Add minidlna plugin
The plugin gives access to all configuration options (except "user to run as" which is set to "volumio") the minidlna.conf file contains. Currently the options of miniDLNA version 1.1.2 are supported.

The plugin features a rescan button that allows to force a rescan of all media directories for new media files.

It has been tested on Raspberry Pi (Volumio 2.389) and x86 (Volumio 2.387 in a VM) but probably also works on other platforms as long as the minidlna package is available via apt-get.