Update index.js
Fix/workaround: The localization files are scanned only when Volumio starts and the plugin is already enabled. Therefore after just installing and enabling the plugin or after booting Volumio with the plugin in disabled state and then enabling it had the effect that localized strings for toasts etc. were not available (toasts did not show up at all then). Now on every start of the plugin the method "self.commandRouter.loadI18nStrings()" gets called which reads the necessary strings.