Use debounce instead of changed GPIO values
When using the previous version of the plugin, I was noticing that the trigger was executed only very rarely.
The reason for this seem was that the value issued from the GPIO was not always `1`. I believe it may have something to do with the button hardware I've been using.
As far as I could see, the motivation for checking the values was only trigger once the value has changed. I can only see a reason in this to debounce.
Thus, I changed the buttons to be debounced by themselves. Now, it works like a charm.

However, I'm not 100% positive about restricting it to `rising`: You need to have a proper pull-uo/down-resistor configured, as far as I can see.