ambx: Fixes and improvement
Fixed a bug where the AmbX plugin could not be uninstalled, if no
AmbX device is connected. It was occuring due to the onStop rejecting
its promise, if the device could not be switched off prior to
uninstalling. It was fixed by onStop always resolving its promise.

Added improvements:
Stability: Color change is now executed with execSync with the
exception onStop. This fixes the lights sometimes not changing color.

Logging: Python logs to syslog and returns error codes.